handler refactor
